refactor:user suggestion component
