+++ src/components/index.js
@@ -7,3 +7,4 @@ export * from "./authComponent/RestrictAuth"
 export * from "./modal/Modal"
 export * from "./comment/SingleComment"
 export * from "./modal/EditCommentModal"
+export {UserList} from "./userslist/UserList"
